CHATBOT-RAGSugg/rag/requirements.txt
# Core dependencies
langchain==0.1.0
chromadb==0.4.15
sentence-transformers==2.2.2

# PDF processing
pypdf==3.17.4

# Vector storage and embeddings
sentence-transformers==2.2.2
torch

# API requests
requests==2.31.0

# Testing
pytest==7.4.3

# AWS (optional)
boto3==1.34.0

# Utilities
numpy
pandas
tqdm

# JSON processing
python-json-logger